requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
schedule==1.2.1
